-Created HeaderFuzzer class in header_fuzzer.py, moved all of the logic into it -Created utils.py containing functions that I felt wouldn't fit well in the HeaderFuzzer class -Created run.py, moved the main function, entry point, and arg parser to it -Cleaned up the 'if not args.all_headers' check in 'if __name__ == '__main__'' -Created setup.py, moved author information to it. Couldn't find author's email so using the contact page link from his home page.
